Changed org weapons to use keybladeID system and summoning should no longer eat items

+++ src/main/java/online/kingdomkeys/kingdomkeys/network/cts/CSSummonKeyblade.java
@@ -1,5 +1,7 @@
 package online.kingdomkeys.kingdomkeys.network.cts;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 import java.util.function.Supplier;
 
@@ -13,6 +15,7 @@ import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.event.entity.item.ItemTossEvent;
@@ -96,21 +99,15 @@ public class CSSummonKeyblade {
 			
 			if(playerData.getActiveDriveForm().equals(Strings.Form_Anti))
 				return;
-			ItemStack orgWeapon = null;
-			
-			if (playerData.getAlignment() != Utils.OrgMember.NONE) {
-				orgWeapon = playerData.getEquippedWeapon().copy();
-			}
-			
-			if(orgWeapon == null) {
-				if(message.alignment != OrgMember.NONE) {
-					orgWeapon = playerData.getEquippedWeapon().copy();
-				}
-			}
 
 			ItemStack heldStack = player.getMainHandItem();
 			ItemStack offHeldStack = player.getOffhandItem();
 			ItemStack chain = playerData.getEquippedKeychain(DriveForm.NONE);
+			boolean useOrg = false;
+			if (playerData.getAlignment() != Utils.OrgMember.NONE) {
+				chain = playerData.getEquippedWeapon().copy();
+				useOrg = true;
+			}
 			ItemStack extraChain = null;
 			if (message.formToSummonFrom != null) {
 				if (!message.formToSummonFrom.equals(DriveForm.NONE)) {
@@ -123,10 +120,10 @@ public class CSSummonKeyblade {
 				}
 			} else {
 				if(playerData.isAbilityEquipped(Strings.synchBlade)) {
-					if(playerData.getAlignment() == OrgMember.NONE || playerData.getEquippedWeapon() != null && playerData.getEquippedWeapon().getItem() instanceof KeybladeItem) {
+					if(playerData.getAlignment() == OrgMember.NONE || playerData.getEquippedWeapon() != null && playerData.getEquippedWeapon().getItem() instanceof KeybladeItem && playerData.getEquippedKeychain(DriveForm.SYNCH_BLADE) != null) {
 						extraChain = playerData.getEquippedKeychain(DriveForm.SYNCH_BLADE);
 					} else {
-						extraChain = orgWeapon.copy();
+						extraChain = chain.copy();
 						for(ItemStack weapon : playerData.getWeaponsUnlocked()) {
 							if(ItemStack.isSameItem(weapon, extraChain)) {
 								extraChain.setTag(weapon.getTag());
@@ -137,14 +134,19 @@ public class CSSummonKeyblade {
 				}
 			
 			}
+			//list of items created so first found keyblade can be removed in order to prevent it from finding the same keyblade twice
+			List<ItemStack> potentialKeyblades = new ArrayList<>(player.getInventory().items); //items = 0-35
+			potentialKeyblades.addAll(player.getInventory().armor); //armor added to keep indexing consistent 36-39
+			potentialKeyblades.addAll(player.getInventory().offhand); //40
+			int slotSummoned = -1;
+			slotSummoned = Utils.findSummoned(potentialKeyblades, chain);
 			//TODO dual wield org weapons
-			int slotSummoned = Utils.findSummoned(player.getInventory(), chain, false);
+			if (slotSummoned != -1) {
+				potentialKeyblades.set(slotSummoned, ItemStack.EMPTY); //set to empty instead of remove to keep indexing consistent
+			}
 			int extraSlotSummoned = -1;
 			if (extraChain != null)
-				extraSlotSummoned = Utils.findSummoned(player.getInventory(), extraChain, false);
-			if (orgWeapon != null) {
-				slotSummoned = Utils.findSummoned(player.getInventory(), orgWeapon, true);
-			}
+				extraSlotSummoned = Utils.findSummoned(potentialKeyblades, extraChain);
 			ItemStack summonedStack = slotSummoned > -1 ? player.getInventory().getItem(slotSummoned) : ItemStack.EMPTY;
 			ItemStack summonedExtraStack = extraSlotSummoned > -1 ? player.getInventory().getItem(extraSlotSummoned) : ItemStack.EMPTY;
 			if (message.forceDesummon) {
@@ -153,15 +155,13 @@ public class CSSummonKeyblade {
 					offHeldStack = summonedExtraStack;
 				}
 			}
-			if ((message.forceDesummon) || (!ItemStack.matches(offHeldStack, ItemStack.EMPTY) && (Utils.hasKeybladeID(offHeldStack)))) {
+			if ((message.forceDesummon) || (!ItemStack.matches(offHeldStack, ItemStack.EMPTY) && ItemStack.matches(offHeldStack, summonedExtraStack) && (Utils.hasKeybladeID(offHeldStack)))) {
 				if (message.forceDesummon || (!ItemStack.matches(heldStack, ItemStack.EMPTY) && (ItemStack.matches(heldStack, summonedStack)))) {
-					if (offHeldStack.getItem() instanceof KeybladeItem) {
-						if (offHeldStack.getTag().getUUID("keybladeID").equals(extraChain.getTag().getUUID("keybladeID"))) {
-							extraChain.setTag(offHeldStack.getTag());
-							playerData.equipKeychain(message.formToSummonFrom, extraChain);
-							player.getInventory().setItem(extraSlotSummoned, ItemStack.EMPTY);
-							player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
-						}
+					if (offHeldStack.getTag().getUUID("keybladeID").equals(extraChain.getTag().getUUID("keybladeID"))) {
+						extraChain.setTag(offHeldStack.getTag());
+						playerData.equipKeychain(message.formToSummonFrom, extraChain);
+						player.getInventory().setItem(extraSlotSummoned, ItemStack.EMPTY);
+						player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
 					}
 				}
 			} else if (extraSlotSummoned > -1) {
@@ -199,29 +199,29 @@ public class CSSummonKeyblade {
 					}
 				}
 			}
-			if ((message.forceDesummon) || (!ItemStack.matches(heldStack, ItemStack.EMPTY) && (Utils.hasKeybladeID(heldStack) || (orgWeapon != null && (heldStack.getItem() instanceof IOrgWeapon || heldStack.getItem() instanceof KeybladeItem))))) {
+			if ((message.forceDesummon) || (!ItemStack.matches(heldStack, ItemStack.EMPTY) && (Utils.hasKeybladeID(heldStack)))) {
 				//DESUMMON
-				if (heldStack.getItem() instanceof KeybladeItem && orgWeapon == null) {
+				if (Utils.hasKeybladeID(heldStack)) {
 					if (heldStack.getTag().getUUID("keybladeID").equals(chain.getTag().getUUID("keybladeID"))) { //Keyblade user
 						chain.setTag(heldStack.getTag());
-						playerData.equipKeychain(DriveForm.NONE, chain);
+						if (useOrg) {
+							Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
+							for(ItemStack weapon : weapons) {
+								if(ItemStack.isSameItem(weapon, heldStack)) {
+									weapon.setTag(heldStack.getTag());
+									break;
+								}
+							}
+							playerData.setWeaponsUnlocked(weapons);
+						} else {
+							playerData.equipKeychain(DriveForm.NONE, chain);
+						}
+						if(playerData.isAbilityEquipped(Strings.synchBlade) && !extraChain.is(Items.AIR)) {
+							player.getInventory().setItem(40, ItemStack.EMPTY);
+						}
 						player.getInventory().setItem(slotSummoned, ItemStack.EMPTY);
 						player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
 					}
-				} else if (heldStack.getItem() instanceof IOrgWeapon || heldStack.getItem() instanceof KeybladeItem) { //Org user
-					Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
-					for(ItemStack weapon : weapons) {
-						if(ItemStack.isSameItem(weapon, heldStack)) {
-							weapon.setTag(heldStack.getTag());
-							break;
-						}
-					}
-					playerData.setWeaponsUnlocked(weapons);
-					player.getInventory().setItem(slotSummoned, ItemStack.EMPTY);
-					if(playerData.isAbilityEquipped(Strings.synchBlade)) {
-						player.getInventory().setItem(40, ItemStack.EMPTY);
-					}
-					player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
 				}
 			} else if (slotSummoned > -1) {
 				//SUMMON FROM ANOTHER SLOT
@@ -230,15 +230,15 @@ public class CSSummonKeyblade {
 				spawnKeybladeParticles(player, InteractionHand.MAIN_HAND);
 
 			} else {
-				if (!ItemStack.matches(chain, ItemStack.EMPTY) || orgWeapon != null) {
+				if (!ItemStack.matches(chain, ItemStack.EMPTY)) {
 					if (ItemStack.matches(heldStack, ItemStack.EMPTY)) {
 						ItemStack keyblade;
-						if (orgWeapon == null) {
+						if (!useOrg) {
 							keyblade = new ItemStack(((IKeychain) chain.getItem()).toSummon());
 							keyblade.setTag(chain.getTag());
 						} else {
 							//Summon org
-							keyblade = orgWeapon;
+							keyblade = chain;
 							Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
 							for(ItemStack weapon : weapons) {
 								if(ItemStack.isSameItem(weapon, keyblade)) {
@@ -249,17 +249,18 @@ public class CSSummonKeyblade {
 
 						}
 						//Summon when keyblade is unsummoned
+						Utils.swapStack(player.getInventory(), player.getInventory().selected, player.getInventory().getFreeSlot());
 						player.getInventory().setItem(player.getInventory().selected, keyblade);
 						player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.summon.get(), SoundSource.MASTER, 1.0f, 1.0f);
 						spawnKeybladeParticles(player, InteractionHand.MAIN_HAND);
 
 					} else if (player.getInventory().getFreeSlot() > -1) {
 						ItemStack keyblade;
-						if (orgWeapon == null) {
+						if (!useOrg) {
 							keyblade = new ItemStack(((IKeychain) chain.getItem()).toSummon());
 							keyblade.setTag(chain.getTag());
 						} else { //Summon org weapon
-							keyblade = orgWeapon;
+							keyblade = chain;
 							Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
 							for(ItemStack weapon : weapons) {
 								if(ItemStack.isSameItem(weapon, keyblade)) {
@@ -311,7 +312,7 @@ public class CSSummonKeyblade {
 					openContainer.slots.forEach(slot -> {
 						ItemStack stack = slot.getItem();
 						if (slot.container != player.getInventory()) {
-							if ((Utils.hasKeybladeID(stack) && stack.getItem() instanceof KeybladeItem) || (playerData.getAlignment() != Utils.OrgMember.NONE) && ((stack.getItem() instanceof IOrgWeapon || (playerData.getEquippedWeapon().getItem() == stack.getItem())))) {
+							if ((Utils.hasKeybladeID(stack))) {
 								slot.set(ItemStack.EMPTY);
 								if (stack.getItem() instanceof IOrgWeapon || (playerData.getAlignment() != Utils.OrgMember.NONE)) {
 									Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
@@ -340,9 +341,9 @@ public class CSSummonKeyblade {
 			IPlayerCapabilities playerData = ModCapabilities.getPlayer(player);
 			if(playerData != null) {
 				if (droppedItem != null) {
-					if (playerData.getEquippedWeapon() != null) {
-						if ((droppedItem.getItem() instanceof IOrgWeapon || (droppedItem.getItem() instanceof KeybladeItem && playerData.getAlignment() != OrgMember.NONE) || (playerData.getEquippedWeapon().getItem() == droppedItem.getItem()))) {
-							if(droppedItem.getItem() instanceof IOrgWeapon || (droppedItem.getItem() instanceof KeybladeItem && playerData.getAlignment() != OrgMember.NONE)) { //If weapon is org weapon || a keyblade && org member
+					if ((Utils.hasKeybladeID(droppedItem))) {
+						if (playerData.getEquippedWeapon() != null) {
+							if ((droppedItem.getItem() instanceof IOrgWeapon || (droppedItem.getItem() instanceof KeybladeItem && playerData.getAlignment() != OrgMember.NONE) || (playerData.getEquippedWeapon().getItem() == droppedItem.getItem()))) {
 								Set<ItemStack> weapons = playerData.getWeaponsUnlocked();
 								for(ItemStack weapon : weapons) {
 									if(ItemStack.isSameItem(weapon, droppedItem)) {
@@ -351,13 +352,9 @@ public class CSSummonKeyblade {
 									}
 								}
 								playerData.setWeaponsUnlocked(weapons);
+
 							}
-							player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
-							event.setCanceled(true);
-							return;
 						}
-					}
-					if ((Utils.hasKeybladeID(droppedItem) && droppedItem.getItem() instanceof KeybladeItem)) {
 						player.level().playSound(null, player.position().x(),player.position().y(),player.position().z(), ModSounds.unsummon.get(), SoundSource.MASTER, 1.0f, 1.0f);
 						event.setCanceled(true);
 					}
